Add UnitTests for ConfigurationCache layer
