+++ Sources/BraintreeCore/BTConfiguration.swift
@@ -21,7 +21,8 @@ import Foundation
     }
     
     /// :nodoc: Timestamp of initialization of each `BTConfiguration` instance
-    let time = Date().timeIntervalSince1970
+    /// Mutable for testing.
+    var time = Date().timeIntervalSince1970
 
     /// :nodoc: This initalizer is exposed for internal Braintree use only. Do not use. It is not covered by Semantic Versioning and may change or be removed at any time.
     ///  Used to initialize a `BTConfiguration`
